Remove CUDA 12.0 and 12.2 unified RAPIDS devcontainers. (#452)
We now require CUDA 12.5+ to build RAPIDS in conda environments, for a
few reasons:
- We only build libraries with CUDA 12.5+ in CI (and test on older
versions via CUDA compatibility), so CUDA 12.0 and 12.2 would only be
tested in this repo's CI
- We need CUDA 12.5+ to ensure alignment between the `conda-forge` and
`nvidia` channels
- We need CUDA 12.4+ to ensure support for GCC 13, which is what we use
in our CUDA 12 conda environments

Therefore, we drop CUDA 12.0 and 12.2 devcontainers.

I dropped the pip devcontainers too, because I don't see a great reason
to keep them if we are deleting the corresponding conda devcontainers.