Support repository dependency_keys in manifest.yaml (#536)
This PR isolates a change from @trxcllnt in #479 so that we can unblock
CI on this repository.

We need this so we can exclude `xgboost` from cuML's dependency list,
since `libxgboost` brings in `librmm` conda packages, and that breaks
devcontainers building everything from source.

Closes #535.

This temporarily breaks cuML devcontainers because it does not include
the cuML test dependencies, but it unblocks cuML CI.

Planned follow-up actions:
- Create a cuML dependency file key called `devcontainers` with
everything except `xgboost`
- Update the `manifest.yaml` to use that file key instead of the current
set of keys (if the cuML PR merges first, we can modify this PR instead
of needing a follow-up)

---------

Co-authored-by: ptaylor <paul.e.taylor@me.com>
Co-authored-by: James Lamb <jaylamb20@gmail.com>