Update RAPIDS containers to Ubuntu 24.04 (#534)
Currently nightly builds fail for cuDF because of the use of `<format>`.
This requires GCC 13 or newer.

The devcontainers are getting the OS default version of GCC (because
RAPIDS devcontainers don't specify a version). The OS is Ubuntu 22.04
which ships with GCC 11. Updating to Ubuntu 24.04 should solve this
problem. Alternatively we can pin the GCC version, but I like being able
to use system-default toolchains where possible.

---------

Co-authored-by: Paul Taylor <178183+trxcllnt@users.noreply.github.com>