Add uninstall command (#632)
We need to uninstall because otherwise the metadata that remains in
place could cause problems. In particular, the ucxx Python package is
problematic because it is the only package in the unified devcontainer
that simultaneously exports a CMake prefix for a vendored library
(libucxx_python.so) and contains actual Python code that loads it. Other
C++ wheels (e.g. libcudf) contain a library and provide an entry point
for loading it, but that function is called by the Python wheels that
depend on it. Since the ucxx wheel does both, in a devcontainer where we
run `clean-ucxx-python`, we cannot run any `build-*-python` commands for
any packages that use `scikit-build-core` afterwards because the build
commands run `pip install`, which will try to load the `cmake.prefix`
entrypoint of all installed packages including ucxx, which will then
attempt to import the package, which in turn will fail for ucxx because
the libucxx_python.so library does not exist.

---------

Co-authored-by: Paul Taylor <178183+trxcllnt@users.noreply.github.com>
Co-authored-by: Bradley Dice <bdice@bradleydice.com>